+++ models/media.js
@@ -23,9 +23,9 @@ const MediaSchema           = new mongoose.Schema({
         username            : { type: String },
     },
     file: {
-        fullPath            : { type: String },
-        fileBase            : { type: String },
+        filePath            : { type: String },
         fileType            : { type: String },
+        fullName            : { type: String },
     },
     title                   : { type: String, trim: true, required: [true, 'is required'] },
     description             : { type: String, trim: true, required: [true, 'is required'] },